Extract out tab logic into a mixin
